Change type
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@2081 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/c64/Banks/ZeroRAMBank.h
@@ -41,7 +41,7 @@ namespace libsidplayfp
 class PLA
 {
 public:
-    virtual void setCpuPort(int state) =0;
+    virtual void setCpuPort(uint8_t state) =0;
     virtual uint8_t getLastReadByte() const =0;
     virtual event_clock_t getPhi2Time() const =0;
